Adding create_inventory.js
- added new inventory classes
- changed the singleHU_spec.js test to include said classes
